# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для работы с системой учета ООО Ромашка
  version: "1.0.0"
  title: RestAPI
consumes:
  - application/json
  - application/xml
  - application/text
produces:
  - application/json
  - application/xml
  - application/text
basePath: "/api"

paths:
  /reports/formats:
    get:
      summary: Получить список форматов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /reports/{category}/{format}:
    get:
      summary: Получить отчёт по заданным параметрам
      operationId: main.get_report
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере

        400:
          description: Ошибка ввода данных
        default:
          description: Неизвестная ошибка
    parameters:
      - name: category
        in: path
        description: Наименование категории данных (nomenclature, group_nomenclature, range, recipe)
        required: true
        type: string
      - name: format
        in: path
        description: Наименование формата отчёта (CSV, MARKDOWN, JSON, XML, RTF)
        required: true
        type: string
  /filter/{category}:
    post:
      summary: Фильтрация данных по заданной категории
      operationId: main.filter_data
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: category
          in: path
          description: Наименование категории данных (nomenclature, group_nomenclature, range, recipe)
          required: true
          type: string
        - name: body
          in: body
          description: Фильтр для данных
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              id:
                type: string
              filter_option:
                type: string
      responses:
        200:
          description: Данные успешно отфильтрованы
        400:
          description: Ошибка ввода данных
        500:
          description: Ошибка на сервере
  /warehouse/transactions:
    post:
      summary: Получить список транзакций по фильтру
      operationId: main.get_warehouse_transactions
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Фильтр для транзакций
          required: true
          schema:
            type: object
            properties:
              warehouse:
                type: object
                properties:
                  name:
                    type: string
                  unique_code:
                    type: string
                  filter_option:
                    type: string
              nomenclature:
                type: object
                properties:
                  name:
                    type: string
                  unique_code:
                    type: string
                  filter_option:
                    type: string
      responses:
        200:
          description: Список транзакций по указанному фильтру
        400:
          description: Данные отсутствуют или ошибка ввода данных
        500:
          description: Ошибка на сервере

  /warehouse/turnover:
    post:
      summary: Получить оборот на складе по фильтру
      operationId: main.get_warehouse_turnover
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Фильтр для расчета оборота склада
          required: true
          schema:
            type: object
            properties:
              warehouse:
                type: object
                description: Фильтры для склада
                properties:
                  name:
                    type: string
                    description: Имя склада
                  unique_code:
                    type: string
                    description: Код склада
                  filter_option:
                    type: string
                    description: Тип фильтрации
              nomenclature:
                type: object
                description: Фильтры для номенклатуры
                properties:
                  name:
                    type: string
                    description: Имя номенклатуры
                  unique_code:
                    type: string
                    description: Код номенклатуры
                  filter_option:
                    type: string
                    description: Тип фильтрации
              start_period:
                type: string
                format: date
                description: Старт периода (YYYY-MM-DD)
              end_period:
                type: string
                format: date
                description: Конец периода (YYYY-MM-DD)
      responses:
        200:
          description: Оборот на складе
        400:
          description: Данные отсутствуют или ошибка ввода данных
        500:
          description: Ошибка на сервере

  /settings/block_period:
    get:
      summary: Получить текущий block_period
      description: Возвращает текущий block_period, установленный в настройках системы
      operationId: main.get_block_period
      produces:
        - application/json
      responses:
        200:
          description: Успешный ответ с текущим block_period
          schema:
            type: object
            properties:
              block_period:
                type: string
                format: date
                example: "YYYY-MM-DD"
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка